% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hearts.R
\name{eveRloving}
\alias{eveRloving}
\title{eveRloving}
\usage{
eveRloving(love = NULL, size = 150, n.hearts = 50,
  colour1 = "hotpink", colour2 = "red4")
}
\arguments{
\item{love}{Character string containing your name and the name of your beloved one}

\item{size}{grid size (of your love)}

\item{n.hearts}{number of hearts within the grid}

\item{colour1, colour2}{hearts colours}
}
\description{
Lets yout spread the love for/of (R) everywhere...
}
\examples{

library(LOLs)

eveRloving()

eveRloving('Phill + Caroline', 200, 100, '#1f3b51', '#aac6dc')

}
