% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{logo}
\alias{logo}
\title{Logo designer}
\usage{
logo(name, text_size = 20, cols = "Blues", additional_text = NULL,
  expr = NULL, expr.index = seq_along(expr) - 1, all_caps = FALSE,
  include_title = TRUE, family = "Palatino", fontface = "bold")
}
\arguments{
\item{name}{Name of you company/institution (char string)}

\item{text_size}{Logo text size}

\item{cols}{Logo colours - either avalid palette name or colour names}

\item{additional_text}{Additional title text}

\item{expr}{Posibility to add expressions to logo}

\item{expr.index}{Expression index (possition)}

\item{all_caps}{LOGICAL}

\item{include_title}{LOGICAL Include the name of your company in logo title}

\item{family}{Font style}

\item{fontface}{Font type}
}
\description{
Simple function developed to desing logos
}
\examples{

library(RFun)

a <- logo(name = 'KVHEM', 
additional_text = 'Katedra vodního hospodářství 
                             a environmentálního modelování')
a + theme(title = element_text(size = 12.5))

b <- logo(name = 'KVHEM', 
          additional_text = 'Katedra vodního hospodářství 
                             \\na environmentálního modelování', 
          include_title = FALSE, 
          family = 'AvantGarde')
b + theme(panel.grid.major = element_blank(),
          panel.grid.minor = element_blank(),
          axis.ticks = element_blank(),
          axis.text = element_blank(),
          axis.line = element_blank(),
          plot.title = element_text(size = 25, 
                                    hjust = 0.25, 
                                    color = 'steelblue4'))

c <- logo(name = 'R-Users Group', 
          text_size = 10, 
          cols = 'Greens', 
          family = 'Courier')
c

x <- logo(name = 'DRUtES', 
          additional_text = '\\nDual Richards Unsaturated Equation Solver', 
          cols = c('royalblue4', 
                   'lightsteelblue1'), 
          expr = expression(integral()[Omega]), 
          text_size = 9.5, 
          include_title = FALSE, 
          fontface = 'bold.italic')
x + theme(panel.grid.major = element_blank(),
          panel.grid.minor = element_blank(),
          axis.ticks = element_blank(),
          axis.text = element_blank(),
          axis.line = element_line(colour = 'royalblue4'))

}
