% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOLs.R
\docType{package}
\name{LOLs-package}
\alias{LOLs-package}
\title{LOLs}
\description{
LOLs is a package that incorporates data manipulation including parametric equations into framework allowing user to push the limits of creativity and imagination.
}
\author{
\strong{Coded and maintained by:} Filip Strnad \email{strnadf@fzp.czu.cz}
}
