% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{matRix}
\alias{matRix}
\title{matRix}
\usage{
matRix(size.x = 150, size.y = 100, n = 120)
}
\arguments{
\item{size.x}{size of x axis}

\item{size.y}{size of y axis}

\item{n}{max number of strings}
}
\description{
Unfortunately, no one can be told what the Matrix is. You have to see it for yourself.
}
\details{
Extention of \href{https://www.r-bloggers.com/matrix-style-screensaver-in-r/}{R-bloggers article}.
}
\examples{

library(RFun)

matRix()

}
