% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mondrian.R
\name{mondRian}
\alias{mondRian}
\title{mondRian}
\usage{
mondRian(initials = "PM", title = FALSE)
}
\arguments{
\item{initials}{Character string of you initials}

\item{title}{LOGICAL Include funny painting title}
}
\description{
Recreates famous Piet Mondrians painting in ggplot2, hopefully there will be a random Mondrian painting generator included in the future.
}
\details{
Piet Mondrian - Composition with Large Red Plane, Yellow, Black, Grey and Blue, 1921
}
\examples{

library(LOLs)

mondRian()

mondRian('FS') + theme_void()

}
