% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFun.R
\docType{package}
\name{RFun-package}
\alias{RFun-package}
\title{RFun}
\description{
RFun is a package that incorporates data manipulation including parametric equations into framework allowing user to push the limits of creativity and imagination.
}
\author{
\strong{Coded and maintained by:} Filip Strnad \email{strnadf@fzp.czu.cz}
}
