% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hearts.R
\name{rotate}
\alias{rotate}
\title{Matrix rotation}
\usage{
rotate(dta, degree)
}
\arguments{
\item{dta}{input data}

\item{degree}{rotation degree}
}
\description{
internal function
}
\examples{

m <- matrix(data = 1:100,
            ncol = 2)

rotate(dta = m,
       degree = 10)

}
\keyword{internal}
