% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{smiley}
\alias{smiley}
\title{Smiley face :)}
\usage{
smiley(joke = F)
}
\arguments{
\item{joke}{LOGICAL Make it all a f-ing joke? (Watchmen comic book reference)}
}
\description{
Smiley face :)
}
\examples{
library(RFun)
smiley() + theme_void()
smiley(joke = T)
}
