% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{eveRloving}
\alias{eveRloving}
\title{Love is everywhere}
\usage{
eveRloving(love = NULL, size = 150, n.hearts = 50, colour1 = "hotpink",
  colour2 = "red4")
}
\arguments{
\item{love}{Character string containing your name and the name of your beloved}

\item{size}{grid size (of your love)}

\item{n.hearts}{number of hearts within the grid}

\item{colour1}{.}

\item{colour2}{.}
}
\value{
.
}
\description{
Let the (R) world know...
}
\examples{
library(RFun)
eveRloving()
eveRloving('Phill + Caroline', 200, 100, '#1f3b51', '#aac6dc')
}
