% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{mondRian}
\alias{mondRian}
\title{Piet Mondrian - Composition with Large Red Plane, Yellow, Black, Grey and Blue, 1921}
\usage{
mondRian(initials = "FS", title = F)
}
\arguments{
\item{initials}{Character string of you initials}

\item{title}{LOGICAL Include funny painting title}
}
\description{
Recreation of famous Piet Mondrians painting in ggplot2, hopefully there will be a random Mondrian painting generator included in the future.
}
\examples{
library(RFun)
mondRian()
mondRian('JRB') + theme_void()
}
