% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batman.R
\name{secRet}
\alias{secRet}
\title{secRet}
\usage{
secRet(name = "Bruce Wayne")
}
\arguments{
\item{name}{Enter your name (character string) to reveal your biggest secret...}
}
\description{
Reveals your biggest secret...
}
\details{
AKA Batman equation in polygons and ggplot2 (extention of \href{https://www.r-bloggers.com/the-batman-equation/}{R-bloggers article}).
}
\examples{

library(RFun)

secRet('Filip Strnad')

}
