% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{secRet}
\alias{secRet}
\title{...}
\usage{
secRet(name = "Yannis")
}
\arguments{
\item{name}{Enter your name (as character string) to reveal your biggest secret...}
}
\description{
Function revealing that you are actually Bruce Wayne using a secret identity... (function returns grob - graphical object)
}
\examples{
library(RFun)
secRet('Bruce Wayne')
}
