% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smiley_face.R
\name{smiley}
\alias{smiley}
\title{smiley}
\usage{
smiley(joke = FALSE)
}
\arguments{
\item{joke}{LOGICAL Make it all a F-ing joke? (Watchmen comic-book reference)}
}
\description{
puts a smile in your plot tab on a rainy day
}
\examples{

library(LOLs)

smiley() + theme_void()

#' smiley(joke = TRUE)

}
